requirements.txt
streamlit == 1.42.2
opencv_python == 4.11.0.86
matplotlib == 3.10.0
easyocr == 1.7.2
numpy~=2.2.3
pillow~=11.1.0